Initial commit. Implemented KNN algorithm with weights.
Added GUI on top of the KNN algorithm and I have finished validation of inputs on everything except the graph (which I am still working on).

I also added a multitude of helper functions such as normalizing the training set as well as the input set.

All work in this project is a manifestion of my own work with help from various web forums and textbooks.