Added  a datagridview to represent the input data. That is a work in progress.
Fixed an oversight on my part with the normalization not being fully utilized so some points weren't being classified (mainly the user input data point).

All of the points are being correctly classified based on their outputs as well.

Tweaked the size of the program (physical boundries). I may consider modifying the GUI further to make a bit more intuitive.