Changing the layout to a more step-based approach. It will better demonstrate and divide the tasks.
So far I have implemented a bare-bones skeleton and reading in from an excel file.

I have changed the way I view the data as well. There are two assumptions for this program:

1) The class is the first value entered and it is a string.
2) The label for the data point is the second value entered and it is a string.

With that said, I store that information in lists this time around so I have more flexibility and I have a better naming scheme for the entire system as well as greater latitude for extensibility later if I so choose.