Completely switched all extension classes to Lists (for extensibility later).
I fixed the normalization class to work with Lists and fixed some minor bugs.

I fixed the Algorithm class to accept Lists (rather than arrays) as inputs for the same reasons above. I also swapped some of the logic around to make it a bit more straight forward. There is still some refactoring that needs to be done on it though.

TODO: Continue step 1 (initial data) validation to make sure all bases are covered. Need to add a catch for if the user has the file open. Display a dialog and then attempt to open it again.