The added the function to read in from a file for the description (first step) text. Then I parse the description file for various keywords (currently): -<strong></strong> -<em></em> -<li><ul></ul></li> -<li><ol></ol></li> -<h2></h2>
I could extend this further and add coloring and several other features but I think that the list I have is sufficient for now. If I find that I need more functionality, I will add onto the class but currently it is fine.

I also added some error handling for reading in the file in case it isn't found. The program will gracefully exit displaying a dialog message indicating the issue.