Fixed the validation problem that was occuring with the attribute inputs. I had to change the validation class removing all of the error clearing and creating a "RemoveErrors" method in order to facilitate the same thing.
The problem that was happening was that if there was an error and then later if there was a valid input, there was an else statement that would remove the errors and it wouldn't do proper validation.

To fix the problem I changed the method like I outlined above as well as called the RemoveErrors method at the beginning of validation so that it will reset the error count regardless of the previous values (we don't care about them if we are validating again).

Now my focus is on displaying the data. I have a few ideas but most involve changing the algorithm class.