Fixed a normalization problem wherein the data was being normalized locally (input apart from training lists). So the normalization wasn't considering the data as a whole and had a chance to produce errors under the right conditions and skew results.
Fixed the two check boxes in the initialize data step.

Started adding the output to the display step but the output doesn't appear to be working properly. I'm going to investigate that. Even with perfect data and a k value of 1, I should be getting one class but keep getting a different one.