I added all of the files used to dyanmically display text as resources so I can access them without having to create them except for the help file case.
The help file will be generated under Resources\knnHelp.CHM where it must stay to be accessible. If it is moved then the file won't display and the user notified.

I also added the Fody package so I could package everything I need into a single click-once application. Verified to be working.