Cleaned up unused files. Also, I tested the plotting with the distance labels and tried every combination I could think of and couldn't recreate the issue. Maybe I was just confused but maybe there is a bug there somewhere that I can't recreate. I'll have to keep an eye on it. The only thing I could think of is if there are overlapping data points and it is actually adding the distance to the k nearest neighbors but not all of them are displaying because some are overlapping.
Changed the error message for a wrong extension. If the user tries to open a file that has an .xlsx extension but isn't truly that file type then I needed a more inclusive error message.

Cleaned up some white space in all of my files and sorted the using statements.

In the mean time, version 1.00 is done. The program is complete in so far as the initial requirements. I still would like to add extra features like picking the best k and running those algorithms. Also adding discrete data would be nice and finding an elogant way of handling that.