Fixed a small display bug.
Added methods for reading in the data from the excel file.

Removed the workbook parameters from the PopulateDataGrid. The parameters were not needed anymore.

Added support for *.XLS files (Excel 2003).

Extended the ChartExtensions class to handle ALL chart related functions. I cleaned up and methodized the class much more effectively. The code is much more readable now.

I went through the entire program and added methods where code was repeated and I cleaned up quite a bit of code.

I changed the display of the help document when the help button was clicked. It will display the help file for the current step that you are on.

TODO: Add the ability to output a text file that will display a TON of information regarding the classification of your point.

TODO: Add support for Libre Calc (*.ODS) files.