Updated the help file to reflect the recent additions.
I added the ability to generate a file which will output the data of the current run. I will expand upon this in the future to give the user different options on what they want to see and how they want to see it.

However, I did run into an unpredictable error. My validation got broken some how and I'm not sure how or why. So in the meantime, I moved all of the validation back into the main class. I will look into this and see if I can find another, more effective, solution to the problem.

I still need to look into stopping the data points from changing colors when the plot button is clicked. I thought I had fixed it but maybe I was mistaken.